feat(services): add heatmap, history, provenance, and topology loaders
