+++ src/qdash/api/lib/copilot_sandbox.py
@@ -1,5 +1,5 @@
 """Compatibility wrapper for shared copilot sandbox helpers."""
 
-from qdash.common.copilot import sandbox as _sandbox
+from qdash.common.copilot import python_sandbox as _sandbox
 
 execute_python_analysis = _sandbox.execute_python_analysis
