+++ src/qdash/common/copilot/formatters/__init__.py
@@ -1,0 +1,1 @@
+"""Formatting helpers for Copilot data presentation."""
