+++ src/qdash/common/copilot/llm_agent.py
@@ -16,13 +16,13 @@ from typing import TYPE_CHECKING, Any, cast
 
 from openai import AsyncOpenAI, BadRequestError
 
-from qdash.common.copilot.analysis import AnalysisResponse, TaskAnalysisContext
+from qdash.common.copilot.analysis_models import AnalysisResponse, TaskAnalysisContext
 
 if TYPE_CHECKING:
-    from qdash.common.copilot.config import CopilotConfig, ModelConfig
-    from qdash.common.copilot.sandbox import SandboxChartSpec, SandboxResult
+    from qdash.common.copilot.python_sandbox import SandboxChartSpec, SandboxResult
+    from qdash.common.copilot.settings import CopilotConfig, ModelConfig
 else:
-    from qdash.common.copilot.config import CopilotConfig, ModelConfig
+    from qdash.common.copilot.settings import CopilotConfig, ModelConfig
 
 logger = logging.getLogger(__name__)
 
@@ -1596,7 +1596,7 @@ def _wrap_tool_executors(
 
     # execute_python_analysis: data_store auto-injected + chart collection
     def python_wrapper(args: dict[str, Any]) -> SandboxResult | dict[str, Any]:
-        from qdash.common.copilot.sandbox import execute_python_analysis
+        from qdash.common.copilot.python_sandbox import execute_python_analysis
 
         result = execute_python_analysis(args["code"], data_store)
         if isinstance(result, dict) and result.get("chart"):
