+++ src/qdash/common/copilot/services/__init__.py
@@ -1,0 +1,1 @@
+"""Service modules for shared Copilot data loading and context assembly."""
