+++ src/qdash/common/copilot/services/heatmap_service.py
