+++ src/qdash/common/copilot/settings.py
